refactor: utils 관련 파일명 수정
