+++ libs/device-utils.ts
